+++ index.ts
@@ -9,7 +9,7 @@ import authRouter from "./router/authRoutes.js";
 import passport from "passport";
 import cookieParser from "cookie-parser";
 import "./utils/auth.js"; 
-import { expressjwt } from "express-jwt";
+import jwtMiddleware from "./middlewares/jwtMiddleware.js";
 
 dotenv.config();
 const port = process.env.PORT || 8000;
@@ -20,26 +20,6 @@ app.use(express.json());
 app.use(cookieParser());
 app.use(passport.initialize());
 
-
-const jwtMiddleware = expressjwt({
-  secret: process.env.SECRET_KEY!,
-  algorithms: ["HS256"],
-  requestProperty: "user",
-  getToken: (req) => {
-    // Check cookie first
-    if (req.cookies && req.cookies.access_token) {
-      return req.cookies.access_token;
-    }
-    // Fallback to Authorization header
-    if (req.headers.authorization && req.headers.authorization.startsWith("Bearer ")) {
-      return req.headers.authorization.split(" ")[1];
-    }
-    return null;
-  }
-}).unless({
-  path: ["/api/user/sign-up", "/api/user/login", "/api/user/refresh-token", "/google/sign-up"
-    ,"/api/auth/google","/api/auth/google/callback", "/",]
-});
 
 app.use(cookieParser());
 app.use(jwtMiddleware);
